requirements.txt
args==0.1.0
click==6.2
click-plugins==1.0.3
cligj==0.4.0
clint==0.5.1
geopy
GeoAlchemy2>=0.2.6
psycopg2>=2.6
pyparsing>=2
python-slugify>=1.2
requests>=2.6
Shapely>=1.5
snuggs>=1.3
SQLAlchemy>=1.0.11
Unidecode>=0.4.19
